requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
psycopg2-binary==2.9.5
Pillow==9.5.0
drf-spectacular==0.26.1
django-environ==0.10.0